fix/moved-assets-to-public
